combined the terminal publisher with the driver ros architecture
